refactor: relocate Background component to modal directory
