+++ apps/client/src/features/modal/Background.tsx
